Make plantuml exporter configurable
Introduce 'generatr.site.exporter' view property to
switch between exporters. The C4 exporter is still the
default one.
